Add -o flag to disable EC-SRP public key being sent in the MT_CPTYPE_PASSSALT control packet
It may be useful to force MD5 for an old device that could fail in the presence
of additional control packet in the authentication phase.

Most of the time the user should not worry about it - we know which algorithm
to use because EC-SRP uses a larger "salt" to hold the server's public key.