Fix use of undirected broadcast in FreeBSD.
When you bind to a specific interface's IP in FreeBSD, you cannot use
INADDR_BROADCAST as your destination.  While the L3 address will be
255.255.255.255, the L2 address will be that of the gateway.  Instead,
use a directed broadcast and set the sockopt IP_ONESBCAST to support
turning the directed broadcast into an undirected broadcast on the given
interface.

Without this fix, MNDP listening works, but actually using MAC-Telnet to
connect to devices does not.